Add retrieval-vs-ranking evaluation harness and topological reranker
Fixes a real bug found along the way: resample_poly returned all-zeros
on int16 PCM in this scipy build (load_audio.py cast to float64 after
resampling instead of before), silently corrupting every non-44.1kHz
source song's fingerprints.

Adds the Phase 0-3 pipeline from the capstone proposal:
- database.query() now returns the full ranked candidate list + offsets,
  not just a hardcoded top-3, so recall@k and retrieval-vs-ranking
  failures can be measured directly.
- build_eval_database.py / evaluate_retrieval.py establish a clean-clip
  baseline (recall@1/5/10) on a 20-song DB.
- evaluate_robustness.py sweeps pitch-shift/time-stretch to show Stage-1
  degrades via ranking failures, not retrieval failures (correct song
  stays in the candidate list, just drops in rank).
- topological_rerank.py reranks Stage-1's top-N candidates using 0-dim
  persistent homology (single-linkage/MST) over time-normalized
  chroma-CQT -- pitch-shift invariant by construction, since a uniform
  chroma-bin rotation is an isometry of the per-sequence point cloud.
- evaluate_rerank.py measures whether reranking actually recovers
  ranking failures vs. leaves them, and whether it ever demotes an
  already-correct Stage-1 top-1.

+++ load_audio.py
@@ -28,11 +28,14 @@ def load_audio(source, *, mic_duration=None, mic_sample_rate=44100):
     rate, data = wavfile.read(source)
     if data.ndim == 2:
         data = data[:, 0]
+    data = data.astype(np.float64)
 
     # Auto-resampling
     if rate != mic_sample_rate:
         print(f"Fixing sample rate: {rate}Hz -> {mic_sample_rate}Hz")
+        # resample_poly silently returns all-zeros on integer PCM input in
+        # this scipy build, so data must already be float before this call.
         data = resample_poly(data, mic_sample_rate, rate)
         rate = mic_sample_rate
 
-    return data.astype(np.float64), rate
+    return data, rate
